% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCT-function.R
\name{MCT}
\alias{MCT}
\title{Multiplication-combination tests for paired data with missing values in
both arms}
\usage{
MCT(
  x,
  y,
  hypothesis = c("h0m", "h0p"),
  alternative = c("two.sided", "less", "greater"),
  nperm = 10000,
  alpha = 0.05
)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values
representing the first components of the pairs.}

\item{y}{a (non-empty) numeric vector of data values
representing the second components of the pairs.}

\item{hypothesis}{a character string indicating which hypothesis is
to be tested, must be either 'h0m' (default) or 'h0p'.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of 'two.sided' (default), 'greater' or 'less'.
You can specify just the initial letter.}

\item{nperm}{The number of permutations used for calculating the permutation
tests. The default option is 10,000.}

\item{alpha}{A number specifying the significance level;
the default is 0.05.}
}
\value{
A \code{MCT} object containing the following components:
\item{Input}{The data input of the user.}
\item{Descriptive}{Some descriptive statistics of the data.
  Displayed are the number of individuals per variable,
  the mean, and variance.}
 \item{MCT}{The value of the test staistics of the completely observed
 and the incompletely observed data and the p-value of the permutation
 procedure of each test, as well as the overall test decision
 of the multiplication combination test.}
}
\description{
The MCT function calculates the multiplication combination tests
for testig H0m or H0p hypothesis for matched pairs with missingness
in both arms. For testing the Behrens-Fisher problem H0p: {p = 1/2},
MCT is based upon permutation versions of the Munzel (1999) (paired)
and Brunner and Munzel (2002) (unpaired) tests. And, For testing
the null hypothesis H0m: {mu1 = mu2}, MCT is based upon the
randomization version of the paired t-test (Konietschke
and Pauly, 2014) and Janssen's permutation version
for the Welch test (Janssen, 1997) for paired and unpaired
data respectively.
}
\references{
Janssen, A. (1997). Studentized permutation tests
for non-iid hypotheses and the generalized Behrens-Fisher problem.
Statistics & probability letters, 36(1), 9-21.


Munzel, U.(1999). Nonparametric methods for paired samples.
Statistica Neerlandica, 53(3), 277-286.



Brunner, E., Munzel, U. (2000). The Nonparametric
Behrens-Fisher Problem: Asymptotic Theory and a Small Sample Approximation.
Biometrical Journal 42, 17 -25.


Konietschke, F., & Pauly, M. (2014). Bootstrapping and permuting paired
t-test type statistics. Statistics and Computing, 24(3), 283-296.


Amro, L., Konietschke, F., and Pauly, M.(2019).
Multiplication-combination tests for incomplete paired data.
Statistics in Medicince, 38(17), 3243-3255.
}
