% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PBT-function.R
\name{PBT}
\alias{PBT}
\title{Parametric bootstrap tests for paired data with missing values
in a single arm}
\usage{
PBT(x, y, nbsp = 1000, alpha = 0.05)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values
representing the first components of the pairs.}

\item{y}{a (non-empty) numeric vector of data values
representing the second components of the pairs.}

\item{nbsp}{The number of bootstraps used for calculating the
bootstrapped tests. The default option is 1000.}

\item{alpha}{A number specifying the significance level;
the default is 0.05.}
}
\value{
A \code{PBT} object containing the following components:
\item{Input}{The data input of the user.}
\item{Descriptive}{Some descriptive statistics of the data.
  Displayed are the number of individuals per variable,
  the mean, and variance.}
 \item{PBT}{The value of the Wald-type statistic, the ANOVA-type statistic,
  and the MATS-type
 staistic and the p-value of the bootstrapped procedure of each test.}
}
\description{
The PBT function calculates the asymptotic model based bootstrap version
 of three different quadratic forms for testing H0m:{mu1=mu2};
 the Wald-type statistic, the ANOVA-type
 statistic, and the MATS-type statistic for paired data with
 missingness in a single arm.
}
\references{
Amro, L., Pauly, M., and Ramosaj, B. (2019). Asymptotic based
bootstrap approach for matched pairs with missingness in a single-arm.
}
