% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewGUI.R
\name{viewGUI}
\alias{viewGUI}
\title{A graphical user interface for the package MissPair}
\usage{
viewGUI()
}
\description{
This function provides a graphical user interface (GUI) for
calculating statistical tests in matched pairs with missing
values in a single arm or both arms.
}
\details{
The function produces a GUI for the calculation of the test statistics.
 Data can be loaded via the "Browse" button. The formula,
 the missing pattern, testing hypothesis, testing method,
 number of permutations (default: 10,000), number of bootstraps
 (default: 1000) and the significance level alpha
 (default: 0.05) need to be specified.
}
