% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WPT-function.R
\name{WPT}
\alias{WPT}
\title{A weighted permutation test for paired data with missing values in both arms}
\usage{
WPT(x, y, alternative = c("two.sided", "less", "greater"),
  nperm = 10000, alpha = 0.05)
}
\arguments{
\item{x}{a (non-empty) numeric vector of data values
representing the first components of the pairs.}

\item{y}{a (non-empty) numeric vector of data values
representing the second components of the pairs.}

\item{alternative}{a character string specifying the alternative hypothesis,
must be one of 'two.sided' (default), 'greater' or 'less'.
You can specify just the initial letter.}

\item{nperm}{The number of permutations used for calculating
the weighted permutation test. The default option is 10,000.}

\item{alpha}{A number specifying the significance level; the default is 0.05.}
}
\value{
A \code{WPT} object containing the following components:
\item{Input}{The data input of the user.}
\item{Descriptive}{Some descriptive statistics of the data.
  Displayed are the number of individuals per variable,
  the mean, and variance.}
 \item{WPT}{The value of the weighted test staistic
  and the p-value of the permutation procedure, as well as the test
  decision of the weighted permutation test.}
}
\description{
The WPT function calculates the weighted studentized permutation test
for testing H0m:{mu1=mu2} for matched pairs with missingness in both arm.
WPT test #' is a combination of the randomization version
of the paired t-test (Konietschke and Pauly, 2014)
and Janssen's permutation version for the Welch test
(Janssen, 1997) for complate and incomplete data respectively.
}
\references{
Janssen, A. (1997). Studentized permutation tests
for non-iid hypotheses and the generalized Behrens-Fisher problem.
Statistics & probability letters, 36(1), 9-21.


Konietschke, F., & Pauly, M. (2014). Bootstrapping and permuting paired
t-test type statistics. Statistics and Computing, 24(3), 283-296.


Amro, L., and Pauly, M. (2017). Permuting incomplete paired
data: a novel exact and asymptotic correct randomization test.
Journal of Statistical Computation and Simulation, 87(6), 1148-1159.
}
